+++ src/domain/services/quiz_manager.py
@@ -2,15 +2,18 @@
 
 
 from datetime import datetime, timezone
+from src.dal.remote.gemini import GeminiClient
+from src.domain.models.input_model import InputModel
 from src.domain.models.topics_model import TopicModel
 from src.dal.remote.factory import AdapterFactory
 from src.dal.local.db_adapter import DBAdapter
 from src.core.logs import error, debug
 
-class PreviewManager:
+class QuizManager:
     def __init__(self):
         self.db_adapter = DBAdapter()
         self.adapters_factory = AdapterFactory()
+        self.gemini_client = GeminiClient()
 
 
     def get_all_sources(self):
@@ -99,3 +102,59 @@ class PreviewManager:
             updated_at=res.get("updated_at", datetime.now(timezone.utc).isoformat()),
             source_name=res.get("source_name"),
         ).to_dict()
+    
+    def get_input(
+        self,
+        item_name: str,
+        input_identification: str,
+        **adapter_kwargs,   # adapter-specific knobs if needed (e.g., time_window, tagged)
+    ) -> dict[str, any]:
+        adapter = self.adapters_factory.get_adapter(item_name)
+        
+        if not adapter:
+            error(f"No adapter found for source: {item_name}")
+            return InputModel(
+                source_name=item_name,
+                item_name=item_name,
+                input_identification=input_identification,
+                input_data=None,
+                updated_at=datetime.now(timezone.utc).isoformat(),
+            ).to_dict()
+
+        res = adapter.get_input(input_identification=input_identification, **adapter_kwargs)
+
+        return InputModel(
+            source_name=item_name,
+            item_name=item_name,
+            input_identification=input_identification,
+            input_data=res.get("input_data", None),
+            updated_at=res.get("updated_at", datetime.now(timezone.utc).isoformat()),
+        ).to_dict()
+    
+    async def generate_context(
+            self,
+            
+            item_name: str,
+            input_data: dict[str, any],
+            *args,
+            **kwargs
+    ) -> dict[str, any]:
+        adapter = self.adapters_factory.get_adapter(item_name)
+
+        if not adapter:
+            error(f"No adapter found for source: {item_name}")
+            return {"error": "No adapter found"}
+        
+        
+        
+        prompt = adapter.generate_context(input_data=input_data, *args, **kwargs)
+
+        
+        response = await self.gemini_client.generate_text(
+            prompt=prompt,
+            system_instruction=adapter.instructions(),
+            response_mime_type="application/json",
+            temperature=0.7,
+        )
+
+        return response